requirements.txt
Faker
mysql-connector-python
pymysql
mysqlclient
PyAutoGUI